requirements.txt
# Physical AI and Humanoid Robotics Textbook
# Content validation and code example dependencies

# Content validation
pyyaml>=6.0
jsonschema>=4.17.0
markdown-it-py>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code examples (MuJoCo and robotics)
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0

# Optional: MuJoCo simulation (requires separate MuJoCo installation)
# mujoco>=3.0.0

# Documentation build (optional)
# mkdocs>=1.5.0
# mkdocs-material>=9.0.0
